test_chirp3_standalone/requirements.txt
google-cloud-speech>=2.34.0
google-cloud-storage>=2.14.0
